Move actual state machine logic out of the node
